Minor Fixes
Made sure path code was located properly in test.lua and checker.lua.
Modified the loader function in imap4/init.lua so that a module is compiled
under the assert function but not defined by calling it.  This way errors can be
properly reported if there are problems in the loadfile step.