Added Command Pipelining Support
The pchecker.lua file is a mail checker that takes advantage of command
pipelining.

All pipeline support was added to imap/imap4.lua.  It is a first cut and does
not handle literals yet.  Methods added to imap4 object:

- startPipeline()
- endPipeline()
- readResponses(lasttag)

To maximize code reuse and facilitate simple switching between pipelining and
non-pipelining, this approach was adopted.  When starting a section for
pipelined commands, call startPipeline().  Then call normal command methods as
usual, the difference is they are being staged and not sent.  Also, instead of
returning a response, they return the {tag, command} pair.  Upon calling
endPipeline(), the commands are sent.

The readResponses(lasttag) method is used to read the pipelined command
responses as they come in.  The "lasttag" parameter is the tag of the last
command in the pipeline.  Without it, the function cannot complete.