Extended readResponses functionality
The new readReponses method can now be used to iterate through any range of
responses.  The parameter 'first_tag', if provided, will cause the index into
the __responses table to be initialized to the value that corresponds to a
response with a tag that matches 'first_tag'.  This allows for replaying of
older responses.
If the parameter is skipped, then operation is as before where the first index
is the first response read from the line.