Restructured response handling
Added the __response_handler method to separate getting the response from
handling the response.  It returns the type of response received and normally
only returns on a tagged completion response from the server.  The 'flags'
argument can be used to modify this behavior- most notably for receiving server
greetings or when receiving responses to pipelined commands.

Also added improved commenting, particularly for the readResponse method, which
also now supports a 0 argument for 'first_tag' to facilitate replaying all
responses from the start of a session.