Better Support of Pipelining
Pipelining of commands now accounts for literals, a max pipeline length and
commands that cause a change in state.  Because this is a single threaded
library the solution in all cases is to automatically send whatever is in the
pipeline but continue to allow the pipeline to built with subsequent commands.

Removed several unnecessary loops during data reception.  This also lead to some
minor restructuring of the code paths for receiving data.

Client state changes are managed in __do_cmd now instead of the entry function.
This changed was made to manage pipelined commands that cause state changes.

Minor string pattern modification in pchecker.lua
Created test3 to test command pipelining more thoroughly.