Performance Improvment
It turns out that pattern matching on successively longer and longer string
results in slower operation.  Who knew?  Luckily, we don't have to search the
pipeline string for a literal pattern, just the most recently appended command.
The mod returned performance so that checker.lua is slower than pchecker.lua, as
it should be.