Added Quoted String Handling and Bug Fix
When building IMAP commands, arguments are searched for ATOM-SPECIAL characters
as defined in RFC3501.  If any are found in the argument, then double quotes are
added to the argument string.  The exceptions are if a string is bracketed by
double quotes of parantheses.

Checker.lua was modified to print a full error response from the server if a
command is rejected.