Bug Fix- APPEND command, FETCH command
imap4/imap4.lua
---------------
The FETCH command supports 3 macros.  These macros are sent as is without the
enclosing '()' as opposed to the data items like ENVELOPE of BODY which must be
enclosed in '()'.  Previously, the distinction was not supported.  Now, it is.

The APPEND command was broken when the __handle_argflag routine was added.  The
change prevented the code from adding the literal argument to the end of
the APPEND, thus the command was rejected by the server.  This is now fixed.

For posterity, I'll note the the problem occurred in the __build_arg_str routine
where the __handle_argflag routine is used.  I modified the code slightly to
make the code flow more understandable given the conditions.

test.lua
--------
Modified to test the newly added macro support in the FETCH command.