First Commit for Asynch Functionality
imap4/imap4.lua
---------------
Added `asynch` option when creating object to set to `true` for asynchronous
operation.

Basic difference is responses are no longer handled by luaimap.  Rather, the
`read` (was `__read`) and `responseHandler` (was `__response_handler`) methods
are used to get a response from the line.  In asynchronous mode, the client is
responsible for handling all data from the server.  No response objects are
returned.

Also added the `getSocket` method to return the socket descriptor currently in
use by the object.

Other miscellaneous enhancements/improvements were made:  when pipelining, only
a state change or literal causes responses to be retrieved in the `__do_cmd`
method; state changes are now managed entirely out of the `responseHandler`
method since all the necessary is available at that time.